gestite Risposte Place

+++ src/main/java/co/develhope/team2/meeplemeet_project_team2/controllers/PlaceController.java
@@ -30,6 +30,9 @@ public class PlaceController {
     @GetMapping("/search/list/{status}")
     public ResponseEntity<Optional<List<Place>>> placeList(@PathVariable String status) {
         Optional<List<Place>> placeList = placeService.getListOfPlaces(status);
+        if(placeList.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(placeList);
     }
 
@@ -37,6 +40,9 @@ public class PlaceController {
     @GetMapping("/search/list/type/{placeType}")
     public ResponseEntity<Optional<List<Place>>> placeTypeList(@PathVariable String placeType) {
         Optional<List<Place>> listPlaces = placeService.getListPlaceType(placeType);
+        if(listPlaces.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(listPlaces);
     }
 
@@ -44,6 +50,9 @@ public class PlaceController {
     @GetMapping("/search/time/now")
     public ResponseEntity<Optional<List<PlaceDTO>>> placeTimeList() {
         Optional<List<PlaceDTO>> listTimePlace = placeService.findOpenPlaceNow();
+        if(listTimePlace.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(listTimePlace);
     }
 
@@ -51,6 +60,9 @@ public class PlaceController {
     @GetMapping("/search/time")
     public ResponseEntity<Optional<List<PlaceDTO>>> placeTimeList2(@RequestParam(name = "at") LocalTime time) {
         Optional<List<PlaceDTO>> listTimePlace2 = placeService.findOpenPlace(time);
+        if(listTimePlace2.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(listTimePlace2);
     }
 
@@ -58,13 +70,19 @@ public class PlaceController {
     @GetMapping("/search/{id}")
     public @ResponseBody ResponseEntity<Optional<Place>> searchPlace(@PathVariable Integer id) {
         Optional<Place> idplace = placeService.getActivePlaceById(id);
-        return ResponseEntity.ok(idplace);
+        if(idplace.isPresent()) {
+            return ResponseEntity.ok(idplace);
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // searches for places based on the name
     @GetMapping("/search/n")
     public ResponseEntity<Optional<List<PlaceDTO>>> searchPlaceByName(@RequestParam(name = "name") String name) {
         Optional<List<PlaceDTO>> searchBy = placeService.getPlaceByName(name);
+        if(searchBy.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(searchBy);
     }
 
@@ -72,37 +90,55 @@ public class PlaceController {
     @GetMapping("/search/a")
     public ResponseEntity<Optional<List<PlaceDTO>>> searchPlaceByAddress(@RequestParam(name = "address") String address) {
         Optional<List<PlaceDTO>> searchBy = placeService.getPlaceByAddress(address);
+        if(searchBy.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(searchBy);
     }
 
     // update information about a place by selecting it with its id
     @PutMapping("/update/{id}")
     public ResponseEntity<Optional<Place>> updatePlace(@PathVariable Integer id, @RequestBody @NotNull Place place) {
-        Optional<Place> updatePlace = placeService.updatePlace(id, place);
-        return ResponseEntity.ok(updatePlace);
+        Optional<Place> getPlaceById = placeService.getPlaceById(id);
+        if(getPlaceById.isPresent()) {
+            Optional<Place> updatePlace = placeService.updatePlace(id, place);
+            return ResponseEntity.ok(updatePlace);
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // reactivation of a deleted place
     @PatchMapping("/reactive/{id}")
     public ResponseEntity<Void> reactivationPlace(@PathVariable Integer id) {
         Optional<Place> idPlace = placeService.getPlaceById(id);
-        placeService.reactivationOfAPlace(id, idPlace.get());
-        return ResponseEntity.ok().build();
+        if(idPlace.isPresent()) {
+            placeService.reactivationOfAPlace(id, idPlace.get());
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // logical delete of a place
     @DeleteMapping("delete/l/{id}")
     public ResponseEntity<Place> deleteLogicalPlace(@PathVariable Integer id) {
-        Optional<Place> idPlace = placeService.getActivePlaceById(id);
-        placeService.deleteLogicalPlace(id, idPlace.get());
-        return ResponseEntity.ok().build();
+        Optional<Place> getPlaceById = placeService.getPlaceById(id);
+        if(getPlaceById.isPresent()) {
+            Optional<Place> idPlace = placeService.getActivePlaceById(id);
+            placeService.deleteLogicalPlace(id, idPlace.get());
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // delete a place by selecting it with its id
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Place> deletePlace(@PathVariable Integer id) {
-        placeService.deletePlaceById(id);
-        return ResponseEntity.ok().build();
+        Optional<Place> getPlaceById = placeService.getPlaceById(id);
+        if(getPlaceById.isPresent()) {
+            placeService.deletePlaceById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     // delete all places
